feat: add Fonts panel for managing custom font families and variants
- Introduced a new FontsPanel component for adding, editing, and deleting custom font families and their variants.
- Updated SandboxLabels to include new labels for font management.
- Enhanced warnings system to notify users of duplicate font variants.
- Modified existing components (CanvasToolbar, HtmlToolbar, PdfToolbar) to support focus and blur events.
- Implemented logic for loading and managing custom font files, including error handling for unsupported formats.
- Updated types to accommodate new font-related functionalities.
