+++ packages/postext-pdf/src/fontCache.ts
@@ -1,6 +1,19 @@
 import type { PDFDocument, PDFFont } from 'pdf-lib';
 import { fontKey, parseFontString } from './fontString';
 
+/** True when `bytes` starts with the `OTTO` magic identifying a CFF-flavored
+ *  OpenType font. These are the ones whose subsetting is pathologically
+ *  slow in pdf-lib, so we embed them whole instead. */
+function isCffOpenType(bytes: Uint8Array): boolean {
+  return (
+    bytes.length >= 4 &&
+    bytes[0] === 0x4f && // O
+    bytes[1] === 0x54 && // T
+    bytes[2] === 0x54 && // T
+    bytes[3] === 0x4f    // O
+  );
+}
+
 export type PdfFontProvider = (
   family: string,
   weight: number,
@@ -37,7 +50,13 @@ export class FontCache {
       Array.from(jobs.entries()).map(async ([key, spec]) => {
         try {
           const bytes = await this.provider(spec.family, spec.weight, spec.style);
-          const font = await this.pdfDoc.embedFont(bytes, { subset: true });
+          // pdf-lib subsets CFF OpenType (.otf, signature `OTTO`) by walking
+          // every glyph at save time, which can take minutes for larger
+          // families. Disable subsetting for CFF — the full face embeds in
+          // seconds, at the cost of a somewhat larger PDF. TrueType fonts
+          // subset normally.
+          const subset = !isCffOpenType(bytes);
+          const font = await this.pdfDoc.embedFont(bytes, { subset });
           this.map.set(key, font);
         } catch {
           this.failed.add(`${spec.family} ${spec.weight}${spec.style === 'italic' ? ' italic' : ''}`);
