+++ packages/postext-sandbox/src/controls/index.ts
@@ -17,6 +17,7 @@ export {
   listCustomFontFamilies,
   getCustomFontFamily,
   isCustomFontFamily,
+  isRemovedCustomFontFamily,
   onCustomFontsChanged,
   missingStandardVariants,
 } from './fontLoader';
