+++ packages/postext-sandbox/src/sidebar/ConfigPanel.tsx
@@ -8,7 +8,6 @@ import { exportConfigToJson, importConfigFromJson } from '../storage/persistence
 import { Tooltip } from '../panels/Tooltip';
 import { ConfirmPopover } from '../panels/ConfirmPopover';
 import { ColorPaletteSection } from './sections/ColorPaletteSection';
-import { CustomFontsSection } from './sections/CustomFontsSection';
 import { PageSection } from './sections/PageSection';
 import { LayoutSection } from './sections/LayoutSection';
 import { BodyTextSection } from './sections/BodyTextSection';
@@ -115,7 +114,6 @@ export function ConfigPanel() {
       </div>
       <div className="min-h-0 flex-1 overflow-y-auto">
         <ColorPaletteSection />
-        <CustomFontsSection />
         <PageSection />
         <LayoutSection />
         <BodyTextSection />
