+++ packages/postext-sandbox/src/sidebar/WarningsPanel.tsx
@@ -12,6 +12,7 @@ function iconFor(kind: WarningPayload['kind']) {
     case 'missingFont':
     case 'missingFontFamily':
     case 'missingFontVariant':
+    case 'duplicateFontVariant':
       return Type;
     case 'looseLine':
       return FileWarning;
@@ -37,6 +38,8 @@ function titleFor(payload: WarningPayload, labels: SandboxLabels): string {
       return labels.warningsMissingFontFamilyTitle;
     case 'missingFontVariant':
       return labels.warningsMissingFontVariantTitle;
+    case 'duplicateFontVariant':
+      return labels.warningsDuplicateFontVariantTitle;
     case 'looseLine':
       return labels.warningsLooseLineTitle;
     case 'headingHierarchy':
@@ -66,6 +69,12 @@ function detailFor(payload: WarningPayload, labels: SandboxLabels): string {
       return `"${payload.family}" — ${labels.warningsMissingFontFamilyDetail}`;
     case 'missingFontVariant':
       return `"${payload.family}" [${formatVariantList(payload.variants)}] — ${labels.warningsMissingFontVariantDetail}`;
+    case 'duplicateFontVariant': {
+      const slots = payload.variants
+        .map((v) => `${v.weight}${v.style === 'italic' ? ' italic' : ''} ×${v.count}`)
+        .join(', ');
+      return `"${payload.family}" [${slots}] — ${labels.warningsDuplicateFontVariantDetail}`;
+    }
     case 'looseLine':
       return `${payload.ratio.toFixed(2)}× · ${labels.warningsThresholdLabel} ${payload.threshold.toFixed(2)}×`;
     case 'headingHierarchy':
@@ -82,7 +91,12 @@ function detailFor(payload: WarningPayload, labels: SandboxLabels): string {
 }
 
 function isFontWarning(kind: WarningPayload['kind']): boolean {
-  return kind === 'missingFont' || kind === 'missingFontFamily' || kind === 'missingFontVariant';
+  return (
+    kind === 'missingFont' ||
+    kind === 'missingFontFamily' ||
+    kind === 'missingFontVariant' ||
+    kind === 'duplicateFontVariant'
+  );
 }
 
 function WarningItem({
@@ -160,9 +174,8 @@ export function WarningsPanel() {
   const handleClick = (w: Warning) => {
     if (isFontWarning(w.payload.kind)) {
       // Surface the custom-font manager so the user can upload the missing
-      // variant or re-add the family. The section also shows the font
-      // picker inputs further down.
-      dispatch({ type: 'SET_PANEL', payload: 'config' });
+      // variant, re-add the family, or disambiguate duplicates.
+      dispatch({ type: 'SET_PANEL', payload: 'fonts' });
       return;
     }
     if (w.sourceStart === undefined) return;
