+++ packages/postext-sandbox/src/types/defaultLabels.ts
@@ -5,6 +5,7 @@ export const DEFAULT_LABELS: SandboxLabels = {
   resources: 'Resources',
   markdownEditor: 'Markdown Editor',
   warnings: 'Warnings',
+  fonts: 'Fonts',
   canvas: 'Canvas',
   html: 'HTML',
   pdf: 'PDF',
@@ -440,6 +441,8 @@ export const DEFAULT_LABELS: SandboxLabels = {
   warningsMissingFontVariantDetail: 'The document uses a weight/style combination that has no uploaded file.',
   warningsMissingFontFamilyTitle: 'Unknown font family',
   warningsMissingFontFamilyDetail: 'Not a Google Font and not uploaded as a custom font.',
+  warningsDuplicateFontVariantTitle: 'Duplicate font variant',
+  warningsDuplicateFontVariantDetail: 'Multiple files share the same weight/style — only one will be used.',
   colorPalette: 'Color palette',
   colorPaletteTooltip: 'Named colors reusable across all color fields. Changing an entry updates every reference.',
   colorPaletteEntryName: 'Name',
