+++ packages/postext-sandbox/src/viewport/pdfFontProvider.ts
@@ -1,6 +1,8 @@
 import type { PdfFontProvider } from 'postext-pdf';
+import type { CustomFontFamily, CustomFontVariant } from 'postext';
 import { decompressWoff2 } from 'postext-pdf';
-import { loadFont } from '../controls/fontLoader';
+import { getCustomFontFamily, loadFont } from '../controls/fontLoader';
+import { getFontFile } from '../storage/fontStorage';
 
 const bytesCache = new Map<string, Promise<Uint8Array>>();
 const weightsCache = new Map<string, Promise<number[] | null>>();
@@ -51,6 +53,49 @@ async function fetchAndDecompress(url: string): Promise<Uint8Array> {
   return decompressWoff2(buf);
 }
 
+/** Pick the best uploaded variant for a requested (weight, style):
+ *  prefer an exact style match by nearest weight; if no same-style variant
+ *  exists, fall back to the nearest-weight variant of the other style. */
+function pickCustomVariant(
+  family: CustomFontFamily,
+  weight: number,
+  style: 'normal' | 'italic',
+): CustomFontVariant | null {
+  if (family.variants.length === 0) return null;
+  const sameStyle = family.variants.filter((v) => v.style === style);
+  const pool = sameStyle.length > 0 ? sameStyle : family.variants;
+  return pool.reduce((best, v) =>
+    Math.abs(v.weight - weight) < Math.abs(best.weight - weight) ? v : best,
+  pool[0]!);
+}
+
+async function loadCustomFontBytes(
+  family: CustomFontFamily,
+  weight: number,
+  style: 'normal' | 'italic',
+): Promise<Uint8Array> {
+  const variant = pickCustomVariant(family, weight, style);
+  if (!variant) {
+    throw new Error(`custom font "${family.name}" has no uploaded variants`);
+  }
+  const file = await getFontFile(variant.fileId);
+  if (!file) {
+    throw new Error(`custom font "${family.name}" is missing its uploaded file`);
+  }
+  const bytes = new Uint8Array(file.buffer);
+  switch (variant.format) {
+    case 'woff2':
+      return decompressWoff2(bytes);
+    case 'ttf':
+    case 'otf':
+      return bytes;
+    case 'woff':
+      throw new Error(
+        `.woff is not supported for PDF rendering; re-upload "${variant.fileName ?? family.name}" as .woff2, .ttf, or .otf`,
+      );
+  }
+}
+
 /**
  * Factory for a `PdfFontProvider` backed by Fontsource's jsdelivr CDN. Google
  * Fonts serves a single variable WOFF2 per family, so pdf-lib would embed
@@ -60,6 +105,15 @@ async function fetchAndDecompress(url: string): Promise<Uint8Array> {
  */
 export function createPdfFontProvider(): PdfFontProvider {
   return async (family, weight, style) => {
+    // Custom families bypass the Google/Fontsource path entirely: resolve
+    // bytes from IndexedDB and only decompress when the uploaded file was
+    // .woff2. Skip the shared `bytesCache` because the user can re-upload
+    // or delete variants at any time, and the fileId captures identity.
+    const custom = getCustomFontFamily(family);
+    if (custom) {
+      return await loadCustomFontBytes(custom, weight, style);
+    }
+
     const key = cacheKey(family, weight, style);
     const cached = bytesCache.get(key);
     if (cached) return cached;
